Row type must read all columns as a MaybeColumn
The 'Row' type must be able to represent any result row and all columns
are "nullable" in Cassandra. Concretely, it is used as the result type
of lightweight transactions in cql-io, which return the full row on a
failed (i.e. conflicting) insert as well as the conflicting values on a
failed update, both of which may contain null values. This problem
showed up on testing the failure cases of lightweight transactions in
cql-io.